requirements.txt
google-generativeai
pinecone
firebase-admin
python-dotenv